Bump pinned default manylinux2010 image version
The current pinned version of the manylinux2010 docker image has an
issue where it is trying to use EoL yum repositories
(see pypa/manylinux#836 for more details). The repository path has been
updated on the latest image. This commit bumps the manylinux2010 docker
image version to use the latest version to get this fix.