Support building from an sdist archive
It is now possible to call the cibuildwheel command line and point it
to a package source distribution archive tarball (sdist) instead of
only a directory. The sdist must be a .tar.gz tarball and will be
extracted in place before being built.

Reference: https://github.com/pypa/cibuildwheel/issues/597
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>