fix: use GNU tar on macOS for OCIContainer interactions
macOS uses BSD tar as the default `tar` command.
This can result in hard to debug issues when building linux wheels on macOS because of incompatibilities between GNU tar (used inside containers) & BSD tar (used on host).
This commit tries to use GNU tar on the host macOS and warns if it is not found.